shiro整合完善

+++ src/main/java/kindle/filter/SecurityFilter.java
@@ -2,9 +2,7 @@ package kindle.filter;
 
 import kindle.mapper.RememberMapper;
 import kindle.pojo.Remember;
-import kindle.pojo.RememberExample;
 import kindle.pojo.User;
-import kindle.repository.RememberRepository;
 import kindle.utils.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.context.support.SpringBeanAutowiringSupport;
@@ -37,9 +35,10 @@ public class SecurityFilter implements Filter {
     @Override
     public void doFilter(ServletRequest req, ServletResponse res, FilterChain chain) throws IOException,
             ServletException {
-
         HttpServletRequest request = (HttpServletRequest) req;
         HttpServletResponse response = (HttpServletResponse) res;
+        String rememberCheckbox = request.getParameter("remember");
+
         User user = (User) request.getSession().getAttribute(Constants.SESSSION_USER_KEY);
         System.out.println(">>>>>>>>>>启动过滤器");
         if (user == null) {
@@ -48,18 +47,20 @@ public class SecurityFilter implements Filter {
                 uuid = URLDecoder.decode(uuid,"utf-8");
                 String[] split = uuid.split(Constants.SEPRETOR_FLAG);
                 String cookieUuid = split[0];
-                Remember remember = rememberMapper.selectOne(new Remember(cookieUuid, null, null));
 
+                Remember remember = rememberMapper.selectByCookUuid(cookieUuid);
+                User dbUser = remember.getUser();
                 if (remember != null) {
-                    request.getSession().setAttribute(Constants.SESSSION_USER_KEY, user); // Login.
+                    request.getSession().setAttribute(Constants.SESSSION_USER_KEY, dbUser); // Login.
                     String invariable_series = remember.getInvariableSeries();
                     String token = PasswordUtils.getMD5(RandomUtils.generateMixString(64));
-                    String key = URLEncoder.encode(invariable_series+Constants.SEPRETOR_FLAG+user.getUserName()
+                    String key = URLEncoder.encode(invariable_series+Constants.SEPRETOR_FLAG+dbUser.getUserName()
                             +Constants.SEPRETOR_FLAG+token, "utf-8");
                     remember.setToken(token);
-                    rememberMapper.insert(remember);
+                    rememberMapper.updateByPrimaryKey(remember);
                     CookieUtils.addCookie(response, Constants.COOKIE_NAME, key, Constants
                             .COOKIE_USERNAME_TIMEOUT);
+                    clearUserCookie(request, response, rememberCheckbox);
                     response.sendRedirect("/index");
                     return;
 
@@ -77,24 +78,28 @@ public class SecurityFilter implements Filter {
             response.sendRedirect("/");
             return;
         } else {
-            String remember = request.getParameter("remember");
-            if (CommonUtils.isEmpty(remember)) {
-                String cookieValue = CookieUtils.getCookieValue(request, Constants.COOKIE_NAME);
-                String[] split = cookieValue.split(Constants.SEPRETOR_FLAG);
-                String cookieUuid = split[0];
-              /*  RememberExample rememberExample = new RememberExample();
-                RememberExample.Criteria rememberCri = rememberExample.createCriteria();
-                rememberCri.andInvariableSeriesEqualTo(cookieUuid);
-                rememberMapper.deleteByExample(rememberExample);*/
-                rememberMapper.delete(new Remember(cookieUuid,null,null));
-                CookieUtils.removeCookie(response, Constants.COOKIE_NAME);
-            }
+            clearUserCookie(request, response, rememberCheckbox);
             response.sendRedirect("/index");
             return;
         }
+    }
 
-
-
+    /**
+     * @description 已记住密码的用户 不再勾选记住我  清除cookie信息
+     * @author hely
+     * @date 2017-09-14
+     * @param
+     */
+    public void clearUserCookie(HttpServletRequest request, HttpServletResponse response, String remember) {
+        if (CommonUtils.isEmpty(remember)) {
+            String cookieValue = CookieUtils.getCookieValue(request, Constants.COOKIE_NAME);
+            if(!CommonUtils.isEmpty(cookieValue)){
+                String[] split = cookieValue.split(Constants.SEPRETOR_FLAG);
+                String cookieUuid = split[0];
+                rememberMapper.delete(new Remember(cookieUuid,null,null));
+                CookieUtils.removeCookie(response, Constants.COOKIE_NAME);
+            }
+        }
     }
 
     @Override
